requirements.txt
 requests==2.31.0
anthropic==0.8.1
requests-oauthlib==1.3.1
python-dotenv==1.0.0
